requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
elastic-transport==8.17.0
elasticsearch==8.17.1
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.11.0
itemloaders==1.3.2
jiter==0.8.2
jmespath==1.0.1
lxml==5.3.0
numpy==2.2.2
openai==1.60.2
packaging==24.2
parsel==1.10.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pluggy==1.5.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pymongo==4.11
pyOpenSSL==25.0.0
pypdfium2==4.30.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
zope.interface==7.2
